# openapidemo/apiserver/src/main/java/com/xyz/openapi/server/model/ContractEmployee.java
package com.xyz.openapi.server.model;

public class ContractEmployee extends AbstractEmployee {
    private String vendorId;
    private String vendorName;

    public String getVendorId() {
        return vendorId;
    }

    public void setVendorId(String vendorId) {
        this.vendorId = vendorId;
    }

    public String getVendorName() {
        return vendorName;
    }

    public void setVendorName(String vendorName) {
        this.vendorName = vendorName;
    }
}