# openapidemo/apiserver/src/main/java/com/xyz/openapi/server/model/Department.java
package com.xyz.openapi.server.model;

public class Department {
    private String id;
    private String name;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}